feat: externalize wagmi default configuration data
