+++ packages/app-portal/src/systems/Settings/providers/ConnectProvider.tsx
@@ -1,22 +1,12 @@
-import {
-  ALCHEMY_ID,
-  INFURA_ID,
-  IS_ETH_DEV_CHAIN,
-  WALLETCONNECT_ID,
-} from 'app-commons';
 import { ConnectKitProvider } from 'connectkit';
 import type { Mode } from 'connectkit/build/types';
 import { useTheme } from 'next-themes';
 import type { ReactNode } from 'react';
-import { type Chain, fallback } from 'viem';
-import {
-  http,
-  type CreateConnectorFn,
-  WagmiProvider,
-  createConfig,
-} from 'wagmi';
-import { coinbaseWallet, injected, walletConnect } from 'wagmi/connectors';
-import { ETH_CHAIN } from '~portal/systems/Chains/config';
+
+import { WagmiProvider, createConfig } from 'wagmi';
+import { CHAINS_TO_CONNECT, TRANSPORTS } from '~portal/systems/Chains';
+
+import { generateETHConnectors } from '~portal/systems/Core/utils/connectors';
 
 declare module 'wagmi' {
   interface Register {
@@ -24,49 +14,10 @@ declare module 'wagmi' {
   }
 }
 
-const app = {
-  name: 'Fuel Bridge',
-  description: 'Bridge assets between Fuel and Other Chains',
-  url: 'https://fuels-portal.vercel.app',
-  icons: ['https://fuels-portal.vercel.app/fuel-logo.svg'],
-};
-
-const chainsToConnect = [ETH_CHAIN] as [Chain, ...Chain[]];
-
-const chainName = ETH_CHAIN?.name.toLowerCase();
-const transports = {
-  [chainsToConnect[0].id]: IS_ETH_DEV_CHAIN
-    ? http()
-    : fallback(
-        [
-          http(`https://eth-${chainName}.g.alchemy.com/v2/${ALCHEMY_ID}`),
-          http(`https://${chainName}.infura.io/v3/${INFURA_ID}`),
-          http(),
-        ],
-        { rank: false },
-      ),
-};
-
-const connectors: Array<CreateConnectorFn> = [
-  injected({
-    shimDisconnect: true,
-    target: 'metaMask',
-  }),
-  coinbaseWallet({ appName: app.name, headlessMode: true }),
-];
-if (WALLETCONNECT_ID) {
-  connectors.push(
-    walletConnect({
-      projectId: WALLETCONNECT_ID,
-      showQrModal: false,
-    }),
-  );
-}
-
 export const config = createConfig({
-  chains: chainsToConnect,
-  connectors,
-  transports,
+  chains: CHAINS_TO_CONNECT,
+  connectors: generateETHConnectors(),
+  transports: TRANSPORTS,
   ssr: true,
 });
